+++ dst/app/middlewares/validator.js
@@ -10,11 +10,6 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.validator = void 0;
-/**
- * バリデーターミドルウェア
- * リクエストのパラメータ(query strings or body parameters)に対するバリデーション
- */
-const domain_1 = require("@chevre/domain");
 const express_validator_1 = require("express-validator");
 const http_status_1 = require("http-status");
 const api_1 = require("../error/api");
@@ -24,7 +19,7 @@ function validator(req, __, next) {
         if (!validatorResult.isEmpty()) {
             const errors = validatorResult.array()
                 .map((mappedRrror) => {
-                return new domain_1.chevre.factory.errors.Argument(mappedRrror.param, mappedRrror.msg);
+                return new req.chevre.factory.errors.Argument(mappedRrror.param, mappedRrror.msg);
             });
             next(new api_1.APIError(http_status_1.BAD_REQUEST, errors));
         }