+++ src/@types/index.d.ts
@@ -2,7 +2,7 @@
  * middlewares/authenticationにて、expressのrequestオブジェクトにAPIユーザー情報を追加している。
  * ユーザーの型をここで定義しています。
  */
-import { chevre } from '@chevre/domain';
+import type { chevre } from '@chevre/domain';
 
 declare global {
     namespace Express {
@@ -12,6 +12,7 @@ declare global {
         export interface Request {
             user: IUser;
             accessToken: string;
+            chevre: typeof chevre;
         }
     }
 }