+++ src/app/middlewares/permitScopes.ts
@@ -1,11 +1,8 @@
 /**
  * スコープ許可ミドルウェア
  */
-import { chevre } from '@chevre/domain';
 import * as createDebug from 'debug';
 import { NextFunction, Request, Response } from 'express';
-
-import { } from '../../@types/index';
 
 const debug = createDebug('pecorino-api:middlewares:permitScopes');
 
@@ -35,7 +32,7 @@ export function permitScopes(permittedScopes: IScope[]) {
         try {
             debug('checking scope requirements...', permittedScopesWithResourceServerIdentifier);
             if (!isScopesPermitted(req.user.scopes, permittedScopesWithResourceServerIdentifier)) {
-                next(new chevre.factory.errors.Forbidden('scope requirements not satisfied'));
+                next(new req.chevre.factory.errors.Forbidden('scope requirements not satisfied'));
             } else {
                 next();
             }