+++ src/app/routes/permits.ts
@@ -1,7 +1,7 @@
 /**
  * 許可証ルーター
  */
-import { chevre } from '@chevre/domain';
+import type { chevre } from '@chevre/domain';
 import * as createDebug from 'debug';
 import { Router } from 'express';
 import { body } from 'express-validator';
@@ -32,13 +32,24 @@ permitsRouter.post(
             .not()
             .isEmpty()
             .isString()
-            .isIn([chevre.factory.product.ProductType.MembershipService, chevre.factory.product.ProductType.PaymentCard])
+            // .isIn([chevre.factory.product.ProductType.MembershipService, chevre.factory.product.ProductType.PaymentCard])
+            .custom((value, { req }) => {
+                if (![
+                    req.chevre.factory.product.ProductType.MembershipService,
+                    req.chevre.factory.product.ProductType.PaymentCard
+
+                ].includes(value)) {
+                    throw new Error('invalid issuedThrough.typeOf');
+                }
+
+                return true;
+            })
     ],
     validator,
     async (req, res, next) => {
         try {
             debug('permits findByIdentifier processing...body:', req.body);
-            const permitRepo = new chevre.repository.Permit(mongoose.connection);
+            const permitRepo = await req.chevre.repository.Permit.createInstance(mongoose.connection);
             const permit = await permitRepo.findByIdentifier(
                 {
                     project: { id: { $eq: <string>req.body.project.id } },
@@ -82,12 +93,23 @@ permitsRouter.post(
             .not()
             .isEmpty()
             .isString()
-            .isIn([chevre.factory.product.ProductType.MembershipService, chevre.factory.product.ProductType.PaymentCard])
+            // .isIn([chevre.factory.product.ProductType.MembershipService, chevre.factory.product.ProductType.PaymentCard])
+            .custom((value, { req }) => {
+                if (![
+                    req.chevre.factory.product.ProductType.MembershipService,
+                    req.chevre.factory.product.ProductType.PaymentCard
+
+                ].includes(value)) {
+                    throw new Error('invalid issuedThrough.typeOf');
+                }
+
+                return true;
+            })
     ],
     validator,
     async (req, res, next) => {
         try {
-            const permitRepo = new chevre.repository.Permit(mongoose.connection);
+            const permitRepo = await req.chevre.repository.Permit.createInstance(mongoose.connection);
             const permit = await permitRepo.findByIdentifierAndAccessCode(
                 {
                     project: { id: { $eq: <string>req.body.project.id } },