+++ src/app/routes/router.ts
@@ -4,15 +4,15 @@
 import * as express from 'express';
 
 import { healthRouter } from './health';
-import { ahRouter } from './_ah';
 
 import { accountTransactionsRouter } from './accountTransactions';
 import { cronRouter } from './cron';
 import { permitsRouter } from './permits';
-import { ssktsMembershipCouponRouter } from './ssktsMembershipCoupon';
-import { ssktsSurfrockRouter } from './ssktsSurfrock';
+// import { ssktsMembershipCouponRouter } from './ssktsMembershipCoupon';
+// import { ssktsSurfrockRouter } from './ssktsSurfrock';
 
 import { authentication } from '../middlewares/authentication';
+import { requireDomain } from '../middlewares/requireDomain';
 
 const router = express.Router();
 
@@ -28,16 +28,29 @@ router.get(
         res.send('hello!');
     }
 );
-router.use('/_ah', ahRouter);
-router.use('/cron', cronRouter);
+router.get(
+    '/_ah/warmup',
+    (__, res, next) => {
+        try {
+            res.send('warmup done!');
+        } catch (error) {
+            next(error);
+        }
+    }
+);
 router.use('/health', healthRouter);
+
+// requireDomain(2023-10-06)
+router.use(requireDomain);
+
+router.use('/cron', cronRouter);
 
 // 認証
 router.use(authentication);
 
 router.use('/accountTransactions', accountTransactionsRouter);
 router.use('/permits', permitsRouter);
-router.use('/ssktsMembershipCoupon', ssktsMembershipCouponRouter);
-router.use('/ssktsSurfrock', ssktsSurfrockRouter);
+// router.use('/ssktsMembershipCoupon', ssktsMembershipCouponRouter); // 廃止(2023-10-06~)
+// router.use('/ssktsSurfrock', ssktsSurfrockRouter); // 廃止(2023-10-06~)
 
 export { router };